requirements.txt
asgiref==3.6.0
attrs==22.2.0
bleach==6.0.0
colorama==0.4.6
coverage==7.2.2
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
django-summernote==0.8.20.0
iniconfig==2.0.0
mock==5.0.2
mysqlclient==2.1.1
packaging==23.0
parameterized==0.9.0
Pillow==9.4.0
pluggy==1.0.0
pytest==7.3.1
pytest-django==4.5.2
python-decouple==3.8
regex==2023.3.23
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
webencodings==0.5.1
